backend/requirements.txt

flask-cors
flask-jwt-extended
pymongo
PyPDF2
groq
python-dotenv
werkzeug